eventsService-Tiny/requirements.txt
fastapi
uvicorn[standard]
tinydb
pydantic
python-multipart
httpx==0.27.2
python-dotenv
